requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
crewai==0.141.0
langgraph==0.5.2
requests==2.32.4
typing-extensions==4.14.1
psycopg2-binary==2.9.9
huggingface-hub==0.23.4
python-dotenv==1.0.1